+++ modernized/app/models/card_xref.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, ForeignKey, Numeric, String
+from sqlalchemy import BigInteger, Column, ForeignKey, String
 
 from app.database import Base
 
@@ -7,5 +7,5 @@ class CardXref(Base):
     __tablename__ = "card_xrefs"
 
     card_num = Column(String(16), primary_key=True)
-    cust_id = Column(Numeric(9))
-    acct_id = Column(Numeric(11), ForeignKey("accounts.acct_id"))
+    cust_id = Column(BigInteger)
+    acct_id = Column(BigInteger, ForeignKey("accounts.acct_id"))
